Implement GCD as Apply effect - Also convert i32 to i64 to prevent overflow in calc

+++ sim/src/actor/calc.rs
@@ -13,12 +13,12 @@ pub enum AttackType {
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/damage-and-healing/#direct-damage-d
 pub fn direct_damage(
     sim: &SimState,
-    potency: i32,
+    potency: i64,
     job: lookup::Job,
     stats: &Stats,
     attack_type: AttackType,
     multipliers: Vec<f64>,
-) -> i32 {
+) -> i64 {
     let fatk = attack_power(job, stats.get(Stat::AttackPower));
     let fdet = determination(stats.get(Stat::Determination));
     // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/damage-and-healing/#direct-damage-d
@@ -44,12 +44,12 @@ pub fn direct_damage(
     // ⌊ ⌊ D × buff_1 ⌋ × buff_2 ⌋
     multipliers
         .iter()
-        .fold(d as f64, |total, multiplier| floor(total * *multiplier, 0)) as i32
+        .fold(d as f64, |total, multiplier| floor(total * *multiplier, 0)) as i64
 }
 
 // Level 80 F(AP)
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/functions/#lv-80-fap
-fn attack_power(job: lookup::Job, ap: i32) -> i32 {
+fn attack_power(job: lookup::Job, ap: i64) -> i64 {
     if job.is_tank() {
         // ⌊ 115 · ( AP - 340 ) / 340 ⌋ + 100
         (115 * (ap - 340) / 340) + 100
@@ -61,7 +61,7 @@ fn attack_power(job: lookup::Job, ap: i32) -> i32 {
 
 // F(DET)
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/functions/#determination-fdet
-fn determination(det: i32) -> i32 {
+fn determination(det: i64) -> i64 {
     // ⌊ 130 · ( DET - LevelMod Lv, Main )/ LevelMod Lv, DIV + 1000 ⌋
     130 * (det - lookup::level_modifiers(lookup::LevelColumn::MAIN))
         / lookup::level_modifiers(lookup::LevelColumn::DIV)
@@ -70,7 +70,7 @@ fn determination(det: i32) -> i32 {
 
 // F(TNC)
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/functions/#tenacity-ftnc
-fn tenacity(tnc: i32) -> i32 {
+fn tenacity(tnc: i64) -> i64 {
     // ⌊ 100 · ( TNC - LevelModLv, SUB )/ LevelModLv, DIV + 1000 ⌋
     100 * (tnc - lookup::level_modifiers(lookup::LevelColumn::SUB))
         / lookup::level_modifiers(lookup::LevelColumn::DIV)
@@ -81,7 +81,7 @@ fn tenacity(tnc: i32) -> i32 {
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/functions/#weapon-damage-fwd
 // Use the WD appropriate for the attack being calculated (eg. Auto-attack = physical damage)
 // All weapons have a Physical and Magical Damage value even though one of them is hidden.
-fn weapon_damage(job: lookup::Job, wd: i32) -> i32 {
+fn weapon_damage(job: lookup::Job, wd: i64) -> i64 {
     // ⌊ ( LevelModLv, MAIN · JobModJob, Attribute / 1000 ) + WD ⌋
     (lookup::level_modifiers(lookup::LevelColumn::MAIN)
         * lookup::job_modifiers(job, job.primary_stat())
@@ -91,7 +91,7 @@ fn weapon_damage(job: lookup::Job, wd: i32) -> i32 {
 
 // P(CHR)
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/parameters/#critical-hit-probability
-fn critical_hit_rate(chr: i32) -> f64 {
+fn critical_hit_rate(chr: i64) -> f64 {
     // ⌊ 200 · ( CHR - LevelModLv, SUB )/ LevelModLv, DIV + 50 ⌋ / 10
     floor(
         200.0 * ((chr - lookup::level_modifiers(lookup::LevelColumn::SUB)) as f64)
@@ -101,13 +101,13 @@ fn critical_hit_rate(chr: i32) -> f64 {
     ) / 10.0
 }
 
-fn is_crit(sim: &SimState, chr: i32) -> bool {
+fn is_crit(sim: &SimState, chr: i64) -> bool {
     let roll = sim.rng.random();
     let probability = critical_hit_rate(chr) / 100.0;
     roll < probability
 }
 
-fn critical_hit_damage(crit: i32) -> i32 {
+fn critical_hit_damage(crit: i64) -> i64 {
     // ⌊ 200 · ( CRIT - LevelModLv, SUB )/ LevelModLv, DIV + 1400 ⌋
     200 * (crit - lookup::level_modifiers(lookup::LevelColumn::SUB))
         / lookup::level_modifiers(lookup::LevelColumn::DIV)
@@ -116,7 +116,7 @@ fn critical_hit_damage(crit: i32) -> i32 {
 
 // F(CRIT)
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/functions/#critical-hit-damage-fcrit
-fn critical_hit(sim: &SimState, crit: i32) -> i32 {
+fn critical_hit(sim: &SimState, crit: i64) -> i64 {
     if !is_crit(sim, crit) {
         return 1000;
     }
@@ -125,7 +125,7 @@ fn critical_hit(sim: &SimState, crit: i32) -> i32 {
 
 // P(DHR)
 // https://www.akhmorning.com/allagan-studies/how-to-be-a-math-wizard/shadowbringers/parameters/#pdhr
-fn direct_hit_rate(dhr: i32) -> f64 {
+fn direct_hit_rate(dhr: i64) -> f64 {
     // ⌊ 550 · ( DHR - LevelModLv, SUB )/ LevelModLv, DIV ⌋ / 10
     floor(
         550.0 * (dhr as f64 - (lookup::level_modifiers(lookup::LevelColumn::SUB)) as f64)
@@ -134,13 +134,13 @@ fn direct_hit_rate(dhr: i32) -> f64 {
     ) / 10.0
 }
 
-fn is_direct(sim: &SimState, dhr: i32) -> bool {
+fn is_direct(sim: &SimState, dhr: i64) -> bool {
     let roll = sim.rng.random();
     let probability = direct_hit_rate(dhr) / 100.0;
     roll < probability
 }
 
-fn direct_hit(sim: &SimState, crit: i32) -> i32 {
+fn direct_hit(sim: &SimState, crit: i64) -> i64 {
     if is_direct(sim, crit) {
         125
     } else {
@@ -215,14 +215,14 @@ mod test {
 
     pub struct FakeRng {
         random_value: f64,
-        random_from_range_value: i32,
+        random_from_range_value: i64,
     }
 
     impl SimRng for FakeRng {
         fn random(&self) -> f64 {
             self.random_value
         }
-        fn random_from_range(&self, _low_inclusive: i32, _high_exclusive: i32) -> i32 {
+        fn random_from_range(&self, _low_inclusive: i64, _high_exclusive: i64) -> i64 {
             self.random_from_range_value
         }
     }
